requirements.txt
PyQt6==6.7.1
openai-whisper
ffmpeg-python==0.2.0
indic-transliteration
